Clear layout manager in CategoryTable
This allows us to delete the getMinimumSize/getMaximumSize overrides.
Previously, the overrides prevented the default FlowLayout from
providing their implementation, which produced nonsense since it's not
actually acting as the layout manager of this class. However, it's
simpler to clear the layout (like JTable does) and allow the
calls to fall through to Component.

Bug: 418717714
Test: CategoryTableTest
Change-Id: I449890ffc9d67c0113291bd584e22add8fc7f0b7

GitOrigin-RevId: a2b700ce06aac3a04dd3097efb62477feae8e6bb