Temporarily convert assert into warning-level log
When a Compose Preview has a Scaffold and potentially other Composables,
we were throwing an Assertion error. That happens because we don't
expect to be in such state. We're temporarily disabling the assertion to
prevent a crash while we investigate the root cause. Once we pinpoint
the issue, we can re-enable the assertion.

Fixes: 418665801
Test: N/A. Manual.
Change-Id: I3b10c6239405f8eb2dfd605132cfc865f50c9f9d

GitOrigin-RevId: b5f2df7461c5a5cea4da5b5d3d2f671fb45d11dd