[Lint,AUA] Make the interaction harder to misuse
Instead of having the interface be three implicitly-coupled methods,
let there instead be two explicitly-coupled methods, where the return
value of one must be used as an argument to the other.

In a somewhat ugly way, hide the `agpVersion` information in an
implementation of the interface; ugly because it requires runtime type
information to retrieve, but nevertheless this at least captures the
functional flow and prevents accidental misuse.

Bug: 421245203
Test: makes me feel better about there not being an automated one
Change-Id: I2e5a466f305fdd8b1fde7e8d13843d1004393466

GitOrigin-RevId: ca9372d63db5b612d1e07ccca991a53ef7afac2c