[spellchecker] extract intellij.spellchecker.xml content module (IJPL-189900)
Classes from com.intellij.spellchecker.xml were located in intellij.xml.impl module (which is loaded by the main classloader of the core plugin) and they depend on classes from intellij.spellchecker. So it is necessary to extract them to a content module to convert 'intellij.spellchecker' to a content module.

GitOrigin-RevId: b3b87d940f00fc2835bdb3d9ffb841489b834430