Change how we setup testFixtures for IDE
There is a property that IDEA introduced `isProductionOnTestDependency` in the IDE dependencies that is used when resolving references in the IDE between production and test sources. This properties serves later when getting the scope of PSIReferences in order to expand the search scope and include the soures of the resolved PsiReference if they are of type TEST.

Also, while resolving the testFixtures dependencies, I came across a bug in AndroidGradleOrderEnumeratorHandlerFactory.shouldProcessDependenciesRecursively() that was returning true, and this seems to have been a change that happened in the middle of some refactoring that switched this property's value, but went uncaught. Initially, this function was supposed to return true only for some very old versions of Gradle( < 2.5).
This handler is very related to how we resolve the test references (i.e. TestFixtures in this case) and would cause us reversing over the testFixtures dependencies and only capturing the main scope (since testFixtures depend on main) instead of what we are actually looking for in the first place.

Bug: 361746109
Test: updated + verified manually for the UI impact
Change-Id: Iaa2b5db4caa68c660cd01cba0691f8f49942cc96

GitOrigin-RevId: f425cf8581ed8a06ff545e377184eefb9aa31559