Fix displaying Gradle Daemons dialog from Memory Settings
The Gradle Daemons was displayed twice when clicking the hyperlink
'Find existing Gradle daemon'. That's because setUI() method is not
called to initialize the view but also to repaint/update it. However,
because that method also invokes addHyperlinkListener linked to a
single instance of myShowDaemonsLabel this end up adding multiple
listeners and as a consequence showing dialog multiple times.

At the same time, continuous loading was displayed when presenting
DaemonsUi view. The reason of it was because the dismiss of loading
happened within invokeLater() which doesn't execute immediately when
a modal dialog is shown. The EDT is blocked while modal dialog is
displayed so only when closing the dialog the code finally invoked.
To address this, a different non-default modality is specified when
creating DaemonUi view.

Fixes: 399814887
Fixes: 368254245
Fixes: 399814711
Fixes: 368246594
Test: Manual
Change-Id: Id018e5569b485ca4c1d316b000001305329fa63a

GitOrigin-RevId: ff19331b62f7a1bb6845cb192489e33681eec097