Convert uses of ProjectSyncModificationTracker...
... to ProjectRootModificationTracker instead in GradleProjectPath

This fixes an issue where if we compute and cache the GradleProjectPath
too early to be null, then it doesn't get invalidated because we use
ProjectRootModificationTracker in findGradleBuildRoot but
ProjectSyncModificationTracker in getGradleBuildIdentityPath.

It also happens to be tested (but not with a proper test), as phased
sync tests intermediately uses the gradle project path to dump tasks.
I'm not sure how we can add a proper test for this.

Bug: 414508111
Fixes: 403303770
Test: Explained above
Change-Id: Id28b8cc2a332f0932474cf92c87b80eb483abc31

GitOrigin-RevId: 3128fea360ff8b15d6d03e4ad1b33f0a66fbbf71