Remove Animation Preview drag support
Dragging animations is not part of our future plans for Animation
Preview, so we should clean up its code to avoid maintainabilty burden.
We're planning to improve the Animation Preview soon, potentially
touching many parts of the code base, so caring about offsets and drag
support should not be a concern going forward.

Bug: N/A
Test: Animation Preview tests
Change-Id: Id5f7223c9c0a2ab28cf7f8250a1f8cb9c3da341a

GitOrigin-RevId: 9210745b444b9560968ae8319e3f8b54ae1422af