Configure imported project with Daemon JVM criteria
Right now, IDE uses a heuristic mechanism to determine which JDK to
use to run Gradle. Incorrect selection may cause the below exception
considering that the fallback in most cases is the Embedded JDK:
- "Unsupported class file major version X" old projects using new JDK
- "UnsupportedClassVersionError" recent projects using old JDK

To improve the existing import mechanism, projects using >=Gradle 9.0
are going to be configured with Daemon JVM criteria using as a version
the highest Java LTS version supported by defined Gradle version.

Given the usage of LTS, the chances of auto-detection mechanism finding
a local installed toolchain are higher, however, in case of failure
auto-provisioning mechanism will download it paying the cost just the
first time in the worst scenario but with a much higher success rate.

Bug: 417959590
Test: ImportProjectWithoutIdeaJdkIntegrationTest
Test: GradleJvmSupportMatrixExtensionsTest
Change-Id: Ic620330f612163384ee3b9dab6f272cc16ec752b

GitOrigin-RevId: 21039dd20d695c4dd1ba8541ba059b482f76fb5a