Extend SafeArgsMode Project System Token
As well as computing the overall mode (Java/Kotlin/None), also allow a
generic computation of the features applicable to this particular
version of SafeArgs generically, rather than implicitly from the
version number of the resolved dependency.  The Gradle implementation
of the token continues to compute the features from the version
number, but other build systems now have the opportunity to behave
appropriately for their own semantics.

This leads to some adjustments to test implementation; we preserve the
test interface (and naming) to be in terms of `Version`, but implement
a conversion to the new set of SafeArgsFeature.

No change in production functionality should be observed.

Bug: 279886738
Test: adjusted existing tests
Change-Id: I49ef45209346fe267779872fe240b31865d3b9bb

GitOrigin-RevId: 34d61500b54000ae0856bf9e3f3cbe39a0431ca0