Prevent invalid values from being saved
It is pretty strange, but the Settings dialog dosn't run validations on apply. The OK and Apply buttons are not disabled when a page contains invalid values and the corresponding boxes have red highlights. It appears that each text field is responsible for not saving its value when it is invalid.

Bug: 422365308
Test: added
Change-Id: I7ed79fabd4096345c3e25bb6f72a69a39a80a81b

GitOrigin-RevId: 1fd628d21fe4e73fe926daeb0c6d559667f4eee9