Add support for empty class and package names in AndroidTestSuiteView
When a test case is added to the `AndroidTestSuiteView` test matrix,
it's nested under a test suite in the test tree. This test suite is
constructed from the test case's `className` and `packageName`.

This change updates the logic to directly add the test case under the
root test suite when both the class and package name are empty strings.
This enables the test suite to report non-class based tests, such as
Journeys.

Bug: 424402224
Test: Added
Change-Id: Icaa92b062bbd3ff5189615db3da4a328dd16cf14

GitOrigin-RevId: af7f23238ac9b26fb702033b912a52bf113aca0d