Strengthen TestProjectSystem a little
Instead of using GradleCoordinate as the backing for both queries and
registrations of dependencies, separate things out into Artifact,
DependencySpec and Dependency data classes.

Make sure that available dependencies are expressed in terms of
Artifacts (with Numeric, non-WILD versions).  Dependency
specifications are expressed in terms of DependencySpec and can be
wild, though since we don't have a concept of "matching" at the moment
that's a little speculative.

Make printing these things look a little bit weird to discourage
anyone from thinking that this tests production behaviour; adjust
tests accordingly.

Bug: 279886738
Test: test-only change
Change-Id: I055b06435e9e3adc811ba90b351f5237105b1842

GitOrigin-RevId: 9779453dbf7ba82d497c670e572ffce7d248cfc2