Set compose jvm property for unit tests
Otherwise, several unit tests receive a MissingResourceException
when creating a compose panel related to this registry flag not being
set.

This is seen with a newer update to Jewel, as this flag is now being
used in JewelComposePanelWrapper. StudioLabs previously did not extend
from IdeaTestSuiteBase so, it is updated to do so here.

Bug: 378750746
Change-Id: Icfdfb5477c9d20a57f8d3fb06a96268642c38b6d

GitOrigin-RevId: 934dfc72c0063bda483741cd2ef49d7c7d17c45c