Record Kotlin Gradle plugin version when compilation crashes (2nd try)
This is a follow-up from previous commit
https://android.googlesource.com/platform/tools/adt/idea/+/035d7c1e48ec2505d26ff67bc9a94d8da6c60e55
submitted in Jun 2024 to record the Kotlin Gradle plugin version being
used when Kotlin compilation crashes.

This version information is contained in an `Exception` subclass named
`KotlinCompilerCrash`.

However, there seems to have been a change in the last year which wraps
`KotlinCompilerCrash` with an outer `Exception` named
`RuntimeExceptionWithAttachments`.

Therefore, this commit will update the extraction logic to get the
Kotlin Gradle plugin version correctly.

Bug: 268521501
Test: Manually verified that the version info is reported correctly
Change-Id: I4d9ce4639785e3e690924256b9cc15e72ba75056

GitOrigin-RevId: c0da6752eae9b4bd963b4362e70359b847ab8aa9