Add "Revert to Original" option in device menu
This change introduces a "Revert to Original" action in the device
selection dropdown menu. This option appears when the preview has been
resized, allowing the user to easily revert the device configuration
back to its original state.

The DeviceChangeListener interface has been updated with an
onRevertToOriginal method to handle this action. The DeviceMenuAction
now checks for a HAS_BEEN_RESIZED data key to determine whether to show
the "Revert to Original" option.

Additionally, a new test has been added to verify that the
onRevertToOriginal listener is correctly triggered when the action is
performed.

Bug: 427284242
Bug: 421128159
Test: RevertToOriginalActionTest
Change-Id: I920207586abc1ba100a6a504dcdfd15ba3dba0d2

GitOrigin-RevId: a6a9e6a05136642124b6d19d2d8e136e5d968e47