Fix devices display in resize panel
The comparator used would return 0 for all values that were not part of
orderOfOption, this would cause only the last item to be present.
This fixes the ordering to make it explicit and ensures that two values
can not result in a 0 comparison since we used also the ordinal to
compare.

Test: CustomConfigurationAttributeCreationPaletteTest
Fixes: 429133510
Change-Id: I1f790dc1f43f0c4464632855200f86cfe5a03522

GitOrigin-RevId: f31628371af99f425e3df2a8e39c769ec92b9c8c