Run GradleProjectPathIntegrationTest phased and non-phased
This is horrible because it turns out we have to override/wrap a
platform service unconditionally in order to be able to turn its
effect on and off with a StudioFlag.

I also don't understand the difference in the actual project path
tests but they are what they are.  (See
I7b20d46ea58e6499b6e946ab113f2ac7c7ba7d14 for the corresponding
change.)

Bug: none filed
Test: this is it
Change-Id: Id9a28b9415349358b1a452392247d375b6310249

GitOrigin-RevId: 1e89d172e42a7a56938289d304f009c9f1ea9c89