De-flake LintIdeTest
The test was not properly waiting on indexing to finish. While this
could have been fixed by adding yet another call to
IndexingTestUtil.waitUntilIndexesAreReady(), we shouldn't have to if
using the platform's test fixtures. This change simplifies the setup.

Note: throughout our codebase, we have many more explicit calls to
wait for indexing than the IntelliJ platform has, which implies we're
still (after 2024.2, that is) missing a few (in our custom test
fixtures) and we have more flakiness as a result.

Fixes: 413743228
Test: LintIdeTest
Change-Id: I1955ea240c45b2149ce7544bcdeec9c893d7b5cc

GitOrigin-RevId: 6a3003c7d034e3e7a5ccd707a2a61deb1bb085e8