Clear errors on device configuration update in ResizePanel
When the device configuration changes in the resize panel,
the width and height text fields should be updated with the
new values.

This change ensures that any prior validation errors on these
fields are cleared, and the new device dimensions are displayed,
as long as the user is not currently editing them.

Fixes: 425242967
Test: manually
Change-Id: I86f19d280d1bbe17abc9f9e3948d9740b83ac100

GitOrigin-RevId: ac025e2dec276f797fdb66879c24c9e3022cc1e0