Separate WorkBench.init for multiple calls
WorkBench.init() was only supposed to be called once.
It looks like we are now calling this from a change listener.
Separate the operations that should only happen once from the
change logic in init().

This fixes a problem where a WorkBench is registered multiple times
with the WorkBenchManager. This caused disposed instances to linger
in the WorkBenchManager.

Fixes: 275377137
Test: WorkBenchTest
Change-Id: I94438438daa486da4b6219bb647ddfd8570c4975

GitOrigin-RevId: e9a5260b5909c08db8c6dd9455c4e346f80ff96d