Fix a test-only utility
AndroidGradleTests#replaceRegexGroup replaces all instances of a
pattern, except if any of them happen to already match the replacement
value, at which point it doesn't replace any subsequent pattern
instances.  This is not what was intended; it was inherited behaviour
from when the replacement only happend on the first instance of any
pattern (when if it matched, the utility could "optimize" by not
changing the input string: see change
Id4e9632724e82e05faf1a7b7978ef0d96eac520a for context).

While we're at it, remove the quadratic scaling behaviour of the
utility.

Remove all the cut-and-pasted instances of the previous test of this
utility in many, many variants of the simpleApplication test project;
test the utility directly.  (Our model snapshot files also effectively
test this, and indeed depend on the correctness in order to be able to
run at all in some cases.)

Bug: none filed
Test: existing
Change-Id: If69b05457439f3b765aa165c87b76c7d976eb566

GitOrigin-RevId: 6730f324ab81b3082d05a4ae38083cca89a4e26a