Make sure we don't store any data in the Sync contributors
Sync contributors should be stateless (extension point), and so having
variables to share state between phases would not respect the context of
 the IDEA project that is being imported (in case of multipl imports at
the same time).
This behavior shows up as a flaky Sync outcome only when multiple
projects are being imported at the same time (hence why failures
weren't captured yet), and is hard to capture in a test environment.

One of the correct ways to store this data would be to use the
ProjectResolverContext as it's a dataholder. This change does that.

Bug: 384022658
Test: verified manually
Change-Id: I3d6b5cc14ff169353c752114c04ce47f1926cd92

GitOrigin-RevId: 7e9b66e0f3f0408d89ab74b055808912afa13c4a