Fix regression of zoom-to-fit when changing tabs
When resetting the zoom-to-fit notifier without waiting for resize only
the NOTIFY_COMPONENT_RESIZED_INT_MASK was applied, this wasn't correct
as if we can resize the component we can suppose the layout was already
created too.

This caused an inconcistency of zoom behaviour when changing Gallery
tabs.

Resetting without waiting for resize caused a race condition since
we were uncorrectly waiting for the layout creation.

To fix this bug it was necessary to also add the newly created mask
NOTIFY_LAYOUT_CREATED_INT_MASK to the reset method

Bug: 383311766
Test: updated
Change-Id: I5d03848e09d6a7b219dcd6b3f2ce8174a9683dc4

GitOrigin-RevId: bfaaffe848df8e56efe8186f8220d3a9032aa626