Adapt screenshot-testing integration test projects
Write the Kotlin Gradle Plugin dependency version in a way that
AndroidGradleTestCase knows how to adapt.  This means that the tests
will run with the latest KGP rather than a hard-coded version, so
remove the test data dependency on the specific KGP-for-Compose
version (which might at some point get updated, so decoupling it from
this specific use is a good thing.

Bug: none filed
Test: existing (test-only change)
Change-Id: I1b4a0496744fbcb6f519e5f3a886a371c8ee7c4e

GitOrigin-RevId: 33bafae9594e29e0810c96e4605aeaf79f5c27f8