Optimize ExpensiveKeepRuleInspection
Currently, this inspection takes a while on medium-size Android
projects, e.g. 60s-90s for nowinandroid or CatchUp.

* the affected classes search should only apply to app modules,
otherwise this is a quadratic operation and since we're looking for
a maximum value, pointless for the other modules.

* we can cut off the search after we hit a certain threshold
* we can filter out the Android SDK entry out of the search scope

With these optimizations it looks like the very broad -keep rules
take <500ms on a Windows desktop to be flagged.

Bug: 421454071
Test: AffectedClassesProjectServiceTest
Change-Id: I3a459de0500f0a8ab5ebb06dcb58893987a7a6d0

GitOrigin-RevId: c920e6201a49afba2f67f6bb2198731d3b552520