Add SettingsSyncTestSuite
This allows for the test suite to be correctly setup by extending
IdeaTestSuiteBase. This is needed for the 2025.2 merge to prevent
classloader exceptions related to "jarFiles is not set for
PluginMainDescriptor".

It remains a bit of a mystery why DateFormatUtil appears to be behaving
differently now on Windows, but this was an intentional decision to use
this API to honor user settings in ag/33464888.

Bug: 431290492
Test: SettingsSyncTestSuite
Change-Id: I8d8af37a1e3936ac4431204722f8fc758ee4147b

GitOrigin-RevId: 8cec9f84e32b437237eedfa0044dadbd6d19b0fe