Deprecate automatic library update for Play Policy Insights
... with the default service compatibility framework.
This commits also adds a studio flag to force Play Policy Insights to
use a specifc version of library. The flag can be used with service
deprecation to resolve unplanned deprecations. We can also use the
flag to prevent studio from updating to the latest versions
temporarily when finding issues in them.

Bug: 421201698
Test: PlayPolicyInsightsJarCacheTest
Change-Id: Id07dce086eeb59c696f95c49507f7fc29529c322

GitOrigin-RevId: 1fa12e3096fe3522ce4accceb6979d9abf4e8e04