Add IDE model for test suites.
Test suites are defined at the project level and are therefore available at the IdeAndroidProject level. The information present at the project level is the junit engine information, the sources and the list of variants the suites apply to.

For Each variant that test suites applies to, there is a test suite object attached which contains the targets for that test suite. Each target will point to the configured target devices and the task name to run the test suite.

Test: TODO
Bug: N/A
Change-Id: I04893b90ead4c354e96d9ad7681af240fd2ee312

GitOrigin-RevId: 3ad8a6007b36882c9fb9655e216c808416298ce8