Fix reference creation on XML attributes
In the case of attributes referencing colors, the attribute can either
be a literal color, e.g. "#ff0000", or a reference to a configuration.

In the case of photos, it can only be a reference to a configuration.

This CL makes changes to allow for reference variants to show up even if
the text does not start with "[". If the text starts with #, then
it's a color literal, and we do not create a reference.

Bug: 392849349
Test: updated UserConfigurationReferenceTest
Change-Id: Ibd3266646beae6f4574fd171de42c214d84a8097

GitOrigin-RevId: 6abdeb07b8c45e402132c3210241c5b06a729616