Remove call to StudioSdkUtil.reloadRemoteSdk()
The caller is already being run under a Task.Backgroundable, but this
method uses StudioProgressRunner to load the repo, which performs a
runWithModalProgressBlocking, which is redundant and blocks the UI.

Use runBlockingCancellable to run the new coroutine-based repo loading
methods under the existing background task.

Update the tests to run the code with a ProgressIndicator in the
context (and replace some mocks with fakes).

Bug: 414350686
Test: ManagedVirtualDeviceCatalogTest
Change-Id: Id23503bd55cf3e7b5bfc90aa3a567d72c0ef25b2

GitOrigin-RevId: 558b4d462279951d211790099f3be5a6689db6e6