Do not delete the Root project moduleEntity in phased sync
This will break the status of Sync (the intermediate one) and we will
end up with Phased sync not actually prividing things early (as
promised). Onbe of these things is the project view that will break
because as we are deleting the Moduleentity at this stage, it happens
after we have already configured the root project entity (as expected)
, its external module options, and its contentRoot entity. All these
entities do share the same moduleEntity object, so things break in a
randon way each time.

Bug: 384022658
Test: verified snapshots content on Android Studio
Change-Id: Iaa5127d1aeae27b608cab60275ca5e642e35518e

GitOrigin-RevId: 0f29cea5ae5085ddb87d224b53c160a6e837cdfb