Prepare migration to new flag storage
This change adds a file that contains the current BooleanFlag
that are not using a static default of false. This also adds
a test that validates the content of this file against the
existing BooleanFlag using reflection.

We want to manage BooleanFlag's default values via a text file
and that will require moving the default values out of the instance
creation constructor calls into the text file. At the same time we will
tweak how values are defined, and this can introduce bugs.

In order to make sure no mistake are made during this mostly manual
process, we first create a file that contains the current values
of these flags and check it in. In order to guarantee this file
is valid, we'll write a test that checks its content.

Bug: N/A
Test: N/A
Change-Id: Ib6e659ebdc00f21883783eb4fd4ff86c6ecbfda0

GitOrigin-RevId: 89a7f9ffb8cb82726a0e8d9d663bae74bdb2a0df