Make `InitAdbLibApplicationServiceRule` configurable
We should make the logic within `InitAdbLibApplicationServiceRule` that
calls `createBridge` on port 5037 configurable. This is crucial because
our unit tests must skip this bridge creation to avoid connecting to
a real ADB server, even if one's available. It's important not to do it
since device tracking would automatically begin and would continue using
default port 5037.

Bug: n/a
Test: existing
Change-Id: Id0667264d4e1caf137048b2215d4f46e4fb080eb

GitOrigin-RevId: 9cf701c1bdf78245c07a0f1565270273817fa901