Extract the table component before doing background work
The ShowMoreElementsItem action is finding the properties table
from the event data. This fails if focus was moved to somewhere
else before clicking on the "Show More" link.

It fails because the table has special logic where the link
component is temporarily added as an editor to handle the click.
After the click is handled the component is removed from the table.

Which is why it may not be accessible through the event data during
the callback from the resource lookup.

Fixes: 435496119
Test: ShowMoreElementsItemTest
Change-Id: I91fd4abf4c25466fe513f1bda9088c2646153d15

GitOrigin-RevId: 83297c27c5cd3ee8425de0631746f170163e04f2