Fix crash in Studio Labs in release builds
The StudioLabsFeaturePanelUi composable was using custom image loading,
directly using `painterResource()`, which does not take the IJP release
classloader quirks, nor any of the other icon loading facilities (e.g.,
picking light/dark variants). The fix is to migrate the UI to use the
Jewel image loading pipeline.

Plus, this fixes the issue of the page not scrolling correctly by
making Compose own the entirety of the page content. This was quite a
trip, since making the Compose content scrollable would cause a native
JBR crash; the solution, for reasons that I am not even 100% sure about
but make intuitive sense, is to mark the configurable as NoScroll. This
makes the Swing container added by the IJP not try to use a viewport,
which seems to be the cause of the native crash if Compose is also doing
the same.

I made the page a lazy grid with an adaptive number of columns (min
width 300 dp); I wanted to make it staggered, but it's not possible yet
due to https://youtrack.jetbrains.com/issue/JEWEL-948

Lastly, this CL removes the unused StudioToolWindowComposePanel API.

Demo: http://go/scrcast/NjI5NjA2NzY2MTM2NTI0OHwxYTQ2NGM0MC00YQ

Fixes: b/435552223
Fixes: b/412732934
Test: verified from ab/ build, maintained existing tests, visual check
Change-Id: I398b77c6b8b68317c704a3eeeeb5b59aae849091

GitOrigin-RevId: 860c0f04b687e23436a507d5854b93768b913872