Check length of attrs before creating a subject
When creating a subject using bazel_rules, it will only check if actual
subject missing some fields compared to the attrs dict we provided.
However, we want to make sure they are the exactly the same. With this
cl, we will first to check the length of actual subject attrs and the
expected attrs are the same. Then we will call the  bazel_rules to
create that subject.

Bug: n/a
Test: existing test
Change-Id: I452ff18406d41254bdc9eb3404442d2d8ddbca85

GitOrigin-RevId: 7dd3e4cbb2079b66803a73b917935f264cd9474d