Fix complication data source reporting it requires a higher
... version than it should.

Complication data sources can have duplicate entries based on the type
of complication. For example, SHORT_TEXT and WEIGHTED_ELEMENT can both
have the COMPLICATION.TEXT data source. However, their required versions
are different. The one in SHORT text requires version 1 and the one in
WEIGHTED_ELEMENT requires version 2.

The code assumed that the entries had the same required version, which
was incorrect. This CL ensures that, for complications, we get the data
source based on the ID but also the complication type.

This CL also removes COMPLICATION_ALL from STATIC_ALL as the data
sources aren't really "static" given they depend on the complication
type.

Fixes: 436552607
Test: added regression tests
Change-Id: Ibe4d14c6e57a84e4589893f003991d23ed2cb9f6

GitOrigin-RevId: d75c3613414f7580ff4fd5eb3a06ce03ea638711