[AUA] Add a threading assertion to ensureParsedModels
It would be nice to do @RequiresBackgroundThread.  Unfortunately, the
way the UpgradeAssistantContentManager tests were written leaked into
the implementation, which does different things in unit test mode, and
tests expect to allow this to run on the EDT.

Do the best we can by asserting the background thread nature in
non-Unit-test mode.

Bug: none filed
Test: existing
Change-Id: I3ee34688c4b31a64651c0ac187187c2b9c571ba9

GitOrigin-RevId: bf8ed19b1895257189660f31634e4216438b7ec3