[AUA,9.0] Fix endpoint for project upgrade integration tests
The transition from before 7.x to 9.x involves two sets of breaking
changes.  The Upgrade Assistant smooths over these changes, but is not
capable of upgrading in one pass.  For our current tests, then,
restrict the domain of target upgrade version to below 9.x in order to
unblock version number bumps; this does unfortunately leave the
integration with 8.x to 9.x upgrades untested.

Bug: 437289201
Test: test-only change
Change-Id: Idacc617d7bf7da5fe0c3b192c70aeedc7fec99b5

GitOrigin-RevId: e561008ef6810de0851a20bbd722813976c3bfc2