[AUA,9.0] Fix endpoint for UI model tests
As with the previous change, it's not in general possible to upgrade
through two major versions of AGP, so we need to ensure that tests of
the UI model don't attempt to do this when we change version numbers.

Inject a "latest known" AGP version explicitly into all the
constructions of the AGP UI models, rather than relying on
AgpVersions.latestKnown.  (We can't override that with a flag because
it only allows upgrades to the "true" latest known AGP version, rather
than downgrades.)

Bug: 437295718
Test: test-only change
Change-Id: Ibd38c78423813707732e4349af7fd8f666bfcaf8

GitOrigin-RevId: 3e67f333d94b203126ac3cf5f163105acb4dcb9d