KMT-1406 [android] (Narwhal Feature Drop Canary 9) Fix test failure after ResourceFolderRepositoryBackgroundActionsTest
There was some code that most likely wasn't using coroutines properly. And most likely they intended to `try {} catch (t: Throwable) { /* ignore */ }`.

But instead they used a supervisorScope, which isn't intended for this.
Now they are blowing exceptions into the `CoroutineExceptionHandler`.
And it seems that this `CoroutineExceptionHandler` in tests has been changed to catch such exceptions in tests more often. (Great change btw).
But this test is now failing.

As a workaround, I fought my urge to refactor the coroutine usage, and instead only override the context for background tasks in the tests.

GitOrigin-RevId: c82667d94a7411120abcdd4630c2f26992f42464