Workaround for IDEA-377644
In IntelliJ 2025.2, certain PsiType subclasses such as PsiArrayType
cache type nullability and do not recompute it when the type
annotations change. We can work around this by constructing a fresh
PsiArrayType instance, forcing nullability to be recomputed.
See https://youtrack.jetbrains.com/issue/IDEA-377644 for details.

Also, this change removes a cast that was not guaranteed to succeed
(since the annotate() method does not always return the same type).

I also added a comment explaining that this util does *not* preserve
existing type annotations (as this is easy to miss).

Bug: 437073861
Test: NullabilityUtilsTest
Test: LightArgsBuilderClassConstructorsAndMethodsTest (with IJ 2025.2)
Change-Id: Idf8ef4bbe066ff37e29343eeab5db9f5aabf2dc7

GitOrigin-RevId: 302eebb3997dfa57ad0a6f310a8739b50113a541