Use IntegerConverter for watch face version property values
When declaring a Watch Face Format version property's value, we expect
an integer. The converter was a ResourceReferenceConverter, which would
offer resource variants when entering a WFF version.

This CL changes the converter for WFF version property values to use
an IntegerConverter instead. For other cases, we continue using a
ResourceReferenceConverter.

This will remove autocompletion for WFF versions, as we will no longer
create resource references.

Fixes: 437043512
Test: PropertyValueConverterTest
Change-Id: I35ae5c3b4277de262188a46574ec13bf208aa503

GitOrigin-RevId: ea658861c54ab639175fa8ed06cca08f176dc50e