Convert most android core gradle tests to JUnit4
These are in submodules of android core for historical reasons, but
this is a convenient grouping to work with in any case.  A few tests
require a little more thought, because of particular things they
expect to be able to do (chiefly, preparing a project without then
subsequently opening it).

This removes the last use of
android/testData/projects/newSyncKotlinTest (which is not the same as
android/testData/snapshots/projects/newSyncKotlinTest), so delete
that.

Bug: none filed
Test: test-only change
Change-Id: I5a711389ceddb0c4158c7068b9fedf0ee0a40c85

GitOrigin-RevId: 85563638bb3deeb4801c0f3de663c41312f0110d