[SafeArgs] Defer registering listeners until needed.
Due to an upstream bug in IJ 2025.2 (more properly, due to a bug in the
bundled version of kotlinc), instances of SafeArgsResolveExtension are
created without being properly disposed via Disposer. This causes
resource leak issues. A fix has been submitted for the upstream bug, but
meanwhile, we can work around the impacts here.

The offending leaked SafeArgsResolveExtension instance is created by the
class KaResolveExtensionToContentScopeRefinerBridge. This instance is
used to look up the extension's shadowed scope (a GlobalSearchScope that
is subtracted from the source scope, to allow the resolve extension to
replace on-disk generated sources). Since we don't actually need any
listeners to compute the shadowed scope, we can defer listener
registration, and therefore participation in the Disposer lifecycle,
until something that requires it is called. This will mean that the
created extensions "leak", but will be harmlessly cleaned up by the GC
just like any other garbage, since they don't have any associated
registered listeners.

Bug: 433681683
Test: included
Change-Id: Ic5e6a9a8a11652d4c3a0d3e1ff4264391d97e073

GitOrigin-RevId: 9fdab18580c58c685bf9e5eeee49823c464f268a