Fix the watch face schema version not updating
The editor was sometimes using the wrong schema version for declarative
watch face files.

When opening Studio, this could happen because the
RawWatchFaceXmlSchemaProvider would be invoked before the merged
manifest had finished computing. The schema would then be considered
"null".

Another situation is when changing the WFF version in the manifest. In
that case, the schema that was previously calculated would be re-used
due to the cache, as the cache for the schema is not based on the PSI
modifier, nor the manifest snapshot. It would only update when changing
the file, as the editor would then drop the cache and re-compute the
schema.

This CL fixes the issues by listening to merged manifest updates and
reparsing declarative watch face files on each update to force the
old schema to be dropped and a new one re-computed.

Fixes: b/437100221
Test: added regression tests
Change-Id: I06d760cb11e32b8d6624a431782edb8a607b5083

GitOrigin-RevId: 73eabcad9c7d2df1f6fcf6affc286055e45aa2f8