Simplify an expression
Some time ago, it was important to make sure that we were being
disciplined about testing AGP models V1 and V2 together for
consistency, not only in the specific version where the default
changed, but also during development of AGP versions which supported
both.  Those times are now long gone, so simplify the expression that
does this and remove the comments which attempt to dissuade the
intrepid maintainer from opting out of exhaustive enumeration checks.

Bug: n/a
Test: test-only change
Change-Id: Ib3f2d14f54bf646cc8d41fe86379d504db86f490

GitOrigin-RevId: b59e5c30ca861ea8bb436cfc927d15088c451bf8