Fix and improve Remove Unused Resources refactor
The Remove Unused Resources refactoring was previously only finding the
default version of each unused resource. For example, if an unused
string resource has many translations, only the default resource would
be removed, breaking the project. Running the refactoring again would
only allow you to remove one translation per run of the refactoring.

We change UnusedResourcesProcessor to use
UnusedResourceDetector.KEY_INCLUDE_ALL_RESOURCE_VERSIONS so that all
resource versions are found for the refactoring. Note that this also
affects the UnusedResourcesQuickFix, so even though the inspection still
only shows the default resource, the quick-fix will in fact remove all
versions of that resource.

We change the filtering logic in UnusedResourcesProcessor. Previously,
if you ran the refactoring while looking at a strings file (where the
strings are unused, and have translations), there would usually be zero
results. This is because we cannot safely remove these strings unless
the translations are included in the scope of the refactoring, so the
refactoring just returns nothing. This was very confusing for the
developer, and made the refactoring seem flaky. We now find the
resources in the current file/scope, and just filter by the resource
names, so the translations are now included in the results, even though
they are outside the currently open file.

It was previously unclear that the refactoring is restricted to the
currently open file or the currently selected files/folders in the
Project tool window. This made the refactoring seem flaky. We add a
"Search entire project" checkbox to the refactoring dialog, and add text
describing what happens if it is left unchecked (the text is changed
depending on whether the focus was on an open editor or not).

Bug: 405676712
Test: updated two existing tests
Change-Id: I9103b7617e7e9acc49408250e5fc2de9bd57b605

GitOrigin-RevId: 9c2501327fc7175715634c6f0d222bf61bba59e3