Remove usage of IntelliJIdeaExternalResourceUrls
IntelliJIdeaExternalResourceUrls is in module
intellij.idea.customization.base which will be removed from
Android Studio when we update to IntelliJ 2025.3.

Since helpPageUrl and keyboardShortcutsPdfUrl point to
JetBrains-hosted content, we are assuming that these URLs remain
stable. This is a reasonable assumption given that JetBrains would not
want to break these links for previous versions of IntelliJ IDEA.

Bug: 430393459
Test: AndroidStudioResourceUrlsTest (updated)
Change-Id: I130da4408c13bf22b278d3ba8dae9fe2af79916f

GitOrigin-RevId: 7d37af3ebe97062c16251c48177925a222603293