Refactor: Remove redundant Compose highlighting logic
This commit removes the redundant color highlighting code for @Composable functions. The functionality is now provided with the bundled Compose Multiplatform (CMP) plugin, which has a higher precedence and was overriding our custom settings. This change streamlines the codebase and resolves a conflict where users were unable to customize the color of composable calls.

The plugin.xml and ComposePluginUtils.kt were modified to remove references to the deleted code and redirect the settings page to the correct IntelliJ attributes.

Test: Verified Manually
Bug: 428296327
Change-Id: Ie3267ce98aba62b92618173024f8741c27f0ddb4

GitOrigin-RevId: a9a7fa924f08551efc3c09d29c8ff3bc3e145122