Add streaming support for non-media audio types
Non-media audio types are supported for API 33+ for 1st party devices and for API 34+ for 3rd party ones.

Non-media audio types are supported by creating an AudioRecord with a custom AudioPolicy. Registration of an AudioPolicy requires the MODIFY_AUDIO_ROUTING permission that was given to the shell account in API 33 on 1st party devices and in API 34 on 3rd party ones.

- An `AudioReader` abstract base class is introduced to encapsulate audio capture logic. This class has two concrete subclasses:
- `AudioRecordReader`: Uses the new AudioRecord/AudioPolicy method.
- `RemoteSubmixReader`: Uses the existing REMOTE_SUBMIX method.

Bug: 406497231
Test: existing
Change-Id: I43bca9a12156b57e03497bfc9e7c25224fcfc809

GitOrigin-RevId: d6100eb05541f2c5cc33d12ef0a68c15a0590f5e