Remove inline diffs from transforms for INTERNAL
We enabled this flag early on in writing controls to help folks try them
out. However, we aren't testing scenarios and are getting some
exceptions that occur because of them.

Since we're about to have a rudimentary version of NEP checked in that
uses these, and since agents already are uses them as well, we don't
need transforms to have a half-baked feature enabled even internally.

If/when in the future we decide we want to use these in-editor diff
controls for transforms for real, we can re-enable these flags once
things fully work.

Bug: n/a
Test: n/a
Change-Id: Id110c95330de3e147080ffbfd698e08b8a4b8250

GitOrigin-RevId: 7600ed813cc3e3a95503ec59b4fbbf47d402cdbf