Make GradleAndroidModelData fully concrete class...
... and use it in data nodes as well.

Required to be able to store it in IntelliJ workspace model

This change is also creating a new class called FileImpl to workaround
the fact that workspace model only accepts data classes. They normally
use VirtualFileUrls to represent files but in our case it would be very
hard to use those. Instead we are using a data class that implements
File that stores the underlying url as a string.

Bug:384022658
Test: covered by existing sync integration tests
Change-Id: I72cc4a274ac66cc133ecfda0a4ab6bbc25695bf5

GitOrigin-RevId: 3190b5cd03bb77f9e02144a11d243e2c0f99a3ea