Update new projects to adopt built in kotlin
With built in kotlin, the kotlinOptions jvmTarget DSL is no longer
present, and is not needed.

This also resolves b/438413026 for the built-in-kotlin case, although it
will still need follow up for the handling of adding new modules to
existing projects that use AGP before 9.0 or have opted-out of
built-in-kotlin.

Bug: 438413026
Bug: 341765853
Test: Included
Change-Id: I19cb8d1a8826ff31449c42e45433b45c5272cf1f

GitOrigin-RevId: ea36aa31479d674c85cfbf59a7f51aba4fb6b889